Move template interface to separate folder
